productCrud

+++ src/main/java/com/isp392/ecommerce/dto/response/ProductResponse.java
@@ -18,5 +18,6 @@ public class ProductResponse {
     String productDescription;
     String status;
     int inStock;
-    BigDecimal price;
+    float price;
+
 }
